[kitsune] Change the argSize option type to uint32_t
For the CPU thread launch runtime functions, this is the size of the argument
bundle. It is pretty standard to use a 32-bit value for this. It is not clear if
there is any reason to have it be larger than this.

+++ kitsune/lib/CodeGen/LowerKitIntrinsics.cpp
@@ -243,7 +243,7 @@ static bool lowerLaunchThreads(CallInst &call, IRBuilder<> &builder) {
 
   Module &m = *getModule(call);
   const DataLayout &dl = m.getDataLayout();
-  uint64_t bundleSize = dl.getTypeStoreSize(bundleTy).getFixedValue();
+  uint32_t bundleSize = dl.getTypeStoreSize(bundleTy).getFixedValue();
   SmallVector<Value *, 4> launchArgs;
   for (unsigned i = 1; i < getNumNonVariadicArgs(call); ++i)
     launchArgs.push_back(call.getArgOperand(i));
