+++ kitsune/runtime/pthreads/kitpthr.h
@@ -80,7 +80,7 @@ void __kitpthr_initialize(void);
 bool __kitpthr_initialized(void);
 KitPthrLaunchContext *__kitpthr_async_launch(KitPthrThrdFunc f, uint64_t start,
                                              uint64_t end, void *args,
-                                             uint64_t argSize);
+                                             uint32_t argSize);
 uint64_t __kitpthr_num_threads(void);
 uint64_t __kitpthr_thread_id(void);
 void __kitpthr_sync(KitPthrLaunchContext *ctx);
